code/requirements.txt
# requirements.txt untuk ESP32 TensorFlow Lite Processor
# Install dengan: pip install -r requirements.txt

# TensorFlow Lite Runtime
tflite-runtime==2.14.0

# Serial communication
pyserial==3.5

# Data processing
numpy==1.24.3

# Optional: untuk logging dan monitoring
matplotlib==3.7.2
pandas==2.0.3
